Solve the prime digit replacmements problem
This was somewhat trivial, since it's just a port of my python code:
https://gist.github.com/Techcable/965341b217ae82defe1f541b3118c328

The original solution was **really** hard,
and took me quite a long time to find a solution.
I believe the solution came to me in the bathtub sometime,
although it may have come to me as I was sitting on a toilet.
This either sends a really deep message about overthinking things,
or says I spend way too much in the restroom.